refactor openstack deploy to use region and move servergroup operations to servergroup packages
