+++ clouddriver-openstack/src/test/groovy/com/netflix/spinnaker/clouddriver/openstack/deploy/ops/servergroup/DeployOpenstackAtomicOperationSpec.groovy
@@ -20,11 +20,10 @@ import com.netflix.spinnaker.clouddriver.data.task.Task
 import com.netflix.spinnaker.clouddriver.data.task.TaskRepository
 import com.netflix.spinnaker.clouddriver.openstack.client.OpenstackClientProvider
 import com.netflix.spinnaker.clouddriver.openstack.client.OpenstackProviderFactory
-import com.netflix.spinnaker.clouddriver.openstack.deploy.description.DeployOpenstackAtomicOperationDescription
+import com.netflix.spinnaker.clouddriver.openstack.deploy.description.servergroup.DeployOpenstackAtomicOperationDescription
 import com.netflix.spinnaker.clouddriver.openstack.deploy.exception.OpenstackOperationException
 import com.netflix.spinnaker.clouddriver.openstack.security.OpenstackNamedAccountCredentials
 import com.netflix.spinnaker.clouddriver.openstack.security.OpenstackCredentials
-import org.openstack4j.api.heat.HeatService
 import spock.lang.Specification
 import spock.lang.Subject
 
@@ -69,8 +68,8 @@ class DeployOpenstackAtomicOperationSpec extends Specification {
     operation.operate([])
 
     then:
-    1 * provider.listStacks() >> []
-    1 * provider.deploy(_ as String, _ as String, _ as Map<String,String>, _ as Boolean, _ as Long)
+    1 * provider.listStacks(_) >> []
+    1 * provider.deploy(_ as String, _ as String, _ as String, _ as Map<String,String>, _ as Boolean, _ as Long)
     noExceptionThrown()
   }
 
@@ -82,8 +81,8 @@ class DeployOpenstackAtomicOperationSpec extends Specification {
     operation.operate([])
 
     then:
-    1 * provider.listStacks() >> []
-    1 * provider.deploy(_ as String, _ as String, _ as Map<String,String>, _ as Boolean, _ as Long) >> { throw new OpenstackOperationException("foobar") }
+    1 * provider.listStacks(_) >> []
+    1 * provider.deploy(_ as String, _ as String, _ as String, _ as Map<String,String>, _ as Boolean, _ as Long) >> { throw new OpenstackOperationException("foobar") }
     OpenstackOperationException ex = thrown(OpenstackOperationException)
     ex.message == "foobar"
   }
